chore: Add priorities for `/usr/bin` and `/bin` directories
